refactor: Extract worker logic into modular handlers and utilities
- packages/proxy/README.md: Add ENVIRONMENT variable documentation and update dev commands
- packages/proxy/package.json: Add --env dev flag to dev scripts
- packages/proxy/src/crypto.ts: Add new signature generation and verification functions
- packages/proxy/src/crypto.test.ts: Add comprehensive tests for crypto functions
- packages/proxy/src/handlers.tsx: Extract route handlers (api/og, home, metadata routes)
- packages/proxy/src/handlers.test.tsx: Add handler integration tests
- packages/proxy/src/og.tsx: Extract OG image generation logic
- packages/proxy/src/rewriter.ts: Extract HTMLRewriter handlers
- packages/proxy/src/utils.ts: Extract utility functions (isDevelopment, isStaticAsset, parsePathMetadata, escapeHtml, getSecret)
- packages/proxy/src/utils.test.ts: Add utility function tests
- packages/proxy/src/worker.tsx: Simplify to routing logic only
- packages/proxy/src/worker.test.tsx: Refactor to test routing instead of implementation details
- packages/proxy/wrangler.toml: Add [env.dev] configuration with ENVIRONMENT variable
