+++ packages/proxy/package.json
@@ -5,8 +5,8 @@
   "main": "dist/worker.js",
   "type": "module",
   "scripts": {
-    "dev": "wrangler dev --config wrangler.toml",
-    "dev:remote": "wrangler dev --remote --config wrangler.toml",
+    "dev": "wrangler dev --config wrangler.toml --env dev",
+    "dev:remote": "wrangler dev --remote --config wrangler.toml --env dev",
     "deploy": "wrangler deploy",
     "build": "tsc",
     "lint": "eslint .",
